pyproject.toml
[tool.poetry]
name = "Elija"
version = "0.1.0"
description = "https://github.com/hpi-sam/ask-your-repository-api"
authors = ["Jascha Beste <jascha.beste@student.hpi.de>"]

[tool.poetry.dependencies]
python = ">=3.7"
Flask = "==1.0.2"
flask-restful = "*"
elasticsearch = "*"
python-dotenv = "*"
gunicorn = "*"
flask-cors = "*"
webargs = "*"
flask-socketio = "*"
eventlet = "*"
requests = "*"
textblob = "*"
nltk = "*"
shovel = "*"
neomodel = "==3.2.9" # version 3.3.0 is broken with flask!

[tool.poetry.dev-dependencies]
coverage = "*"
python-coveralls = "*"
prospector = "^1.1"
mamba = "*"
expects = "*"
doublex = "*"
doublex-expects = "*"
pep8 = "*"
autopep8 = "*"
rope = "*"
PyHamcrest = "*"